% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decapsulate_ml_kem.R
\name{decapsulate_ml_kem}
\alias{decapsulate_ml_kem}
\title{Shared Secret - Retrieve ML-KEM (FIPS 203)}
\usage{
decapsulate_ml_kem(encapsulation, private_key)
}
\arguments{
\item{encapsulation}{Received encapsulation object created by \code{encapsulate_ml_kem()}}

\item{private_key}{Receiver's private key as generated by \code{keygen_ml_kem()}.}
}
\value{
The shared secret
}
\description{
Uses own private key to retrieve the shared secret key created by \code{encapsulate_ml_kem()}.
}
\examples{
key <- keygen_ml_kem(512)
ss1 <- encapsulate_ml_kem(key$public)
ss2 <- decapsulate_ml_kem(ss1$encapsulation, key$private)
identical(ss1$shared_secret, ss2)

}
