% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encapsulate_ml_kem.R
\name{encapsulate_ml_kem}
\alias{encapsulate_ml_kem}
\title{Shared Secret - Create ML-KEM (FIPS 203)}
\usage{
encapsulate_ml_kem(pub_key)
}
\arguments{
\item{pub_key}{Receiver's public key as generated by \code{keygen_ml_kem()}.}
}
\value{
A list of two elements with a secret key and its encapsulation to be shared with the
owner of the used public key.
}
\description{
Creates a shared secret key and encapsulates it for secure transmission.
}
\examples{
key <- keygen_ml_kem(512)
ss <- encapsulate_ml_kem(key$public)  # This generates an encapsulation readable only by key$private

}
