% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope_create.R
\name{envelope_create}
\alias{envelope_create}
\title{Envelope - Create}
\usage{
envelope_create(message, public_key)
}
\arguments{
\item{message}{Message to be encrypted and sent.}

\item{public_key}{The public key produced by \code{keygen_ml_kem()} used to encapsulate the encryption
key used to encrypt the message.}
}
\value{
Invisibly, an envelope with encrypted message and encapsulated key.
}
\description{
Envelope - Create
}
\examples{
key <- keygen_ml_kem(512)
env <- envelope_create("Very important message.", key$public)
}
