% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope_open.R
\name{envelope_open}
\alias{envelope_open}
\title{Envelope - Open}
\usage{
envelope_open(envelope, private_key)
}
\arguments{
\item{envelope}{An envelope produced by \code{envelope_create()}.}

\item{private_key}{A private key able to decapsulate the encryption key contained in the
envelope. It must be the private key paired with the public key used during envelope creation.}
}
\value{
The message.
}
\description{
Envelope - Open
}
\examples{
key <- keygen_ml_kem(512)
env <- envelope_create("Very important message.", key$public)

msg <- envelope_open(env, key$private)
print(msg)
}
