% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keygen_dilithium.R
\name{keygen_dilithium}
\alias{keygen_dilithium}
\title{Dilithium Key Generation}
\usage{
keygen_dilithium(version = 2)
}
\arguments{
\item{version}{Type of key to be generated. Use 2 for Dilithium2, 3 for Dilithium3
or 5 for Dilithium5.}
}
\value{
A \code{keypair} object.
}
\description{
Dilithium Key Generation
}
\examples{
key <- keygen_dilithium()
key$key_type

}
