% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keygen_ml_dsa.R
\name{keygen_ml_dsa}
\alias{keygen_ml_dsa}
\title{Key-Pair Generation - ML-DSA (FIPS 204)}
\usage{
keygen_ml_dsa(strength = 3)
}
\arguments{
\item{strength}{Type of key strength to be generated.
Different strengths are available. Use 2 for ML-DSA-44 (claimed strength
category 2), 3 for ML-DSA-65 (default, claimed strength category 3) or 5
for ML-DSA-87 (claimed strength category 5).
For more information about strength categories see the vignette.}
}
\value{
A \code{keypair} object.
}
\description{
On 13/Aug/2024 NIST published the final version of the Module-Latice-Based
Digital Signature Standard (ML-DSA), based on Crystals-Dilithium submission.
This function generates a keypair to be used by the proposed Digital
Signature Algorithm.
}
\examples{
key <- keygen_ml_dsa()
print(key)

}
\seealso{
https://csrc.nist.gov/pubs/fips/204/final
}
