% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keygen_slh_dsa.R
\name{keygen_slh_dsa}
\alias{keygen_slh_dsa}
\title{Key-Pair Generation - SLH-DSA (FIPS 2025)}
\usage{
keygen_slh_dsa(hash_type = "shake", category = 192, type = "fast")
}
\arguments{
\item{hash_type}{Type of hash to use. Accepts 'shake' (default) or 'sha2'.}

\item{category}{Security category: 128 (claimed security category 1), 192
(default, claimed security category 2) and 256 (claimed security category
3). For more details regarding security categories please refer to the
vignette.}

\item{type}{Type of signature to produce: 'fast' (but having a larger
size) or 'small' (but being slower to compute).}
}
\value{
A \code{pqcrypto_keypair} object.
}
\description{
On 13/Aug/2024 NIST published the final version of the Stateless-Hash-Based
Digital Signature Standard (SLH-DSA), based on Sphincs+ submission.
This function generates a keypair to be used by the proposed Digital
Signature Algorithm.
}
\examples{
# Generate key with default parameters
key1 <- keygen_slh_dsa()
key1$algorithm
key1$param$hash

# Generate key with custom parameters
key2 <- keygen_slh_dsa("sha2", 128, "small")
key2$algorithm
key2$param$hash

}
\seealso{
https://csrc.nist.gov/pubs/fips/205/final
}
