% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_open.R
\name{open_key}
\alias{open_key}
\title{Key Management - Open}
\usage{
open_key(file_name, password = NULL)
}
\arguments{
\item{file_name}{The file containing a key to be opened.}

\item{password}{If the \code{file_name}contains an encrypted private key, the decryption password.}
}
\value{
The public or private key.
}
\description{
Opens a private key or a public key file generated by \code{write_key()}.
}
\examples{
key <- keygen_kyber()
path <- tempdir()
write_key(key$private, path, "myPassword")
prv <- open_key(paste0(path, "/keypair"), "myPassword")

identical(key$private, prv)

}
