% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope_read.R
\name{read_envelope}
\alias{read_envelope}
\title{Envelope - Read}
\usage{
read_envelope(file_name = NULL)
}
\arguments{
\item{file_name}{The file name to read.}
}
\value{
An envelope read from the file
}
\description{
Reads an envelope from a binary file created by \code{write_envelope()}.
}
\examples{
key <- keygen_ml_kem(512)
env <- envelope_create("Very important message.", key$public)
fn <- write_envelope(env)

env2 <- read_envelope(fn)
identical(env, env2)

}
