% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_read.R
\name{read_signature}
\alias{read_signature}
\title{Digital Signature - Read}
\usage{
read_signature(file_name = NULL)
}
\arguments{
\item{file_name}{The file name to read.}
}
\value{
a Digital signature read from the file
}
\description{
Reads a digital signature from a binary file created by \code{write_signature()}.
}
\examples{
key <- keygen_dilithium(2)
important_message <- "Hello world!!"
signature <- sign_dilithium(key$private, important_message)
fname <- write_signature(signature)

signature2 <- read_signature(fname)
identical(signature, signature2)

}
