% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_dillitium.R
\name{sign_dilithium}
\alias{sign_dilithium}
\title{Dilithium Digital Signature - Sign}
\usage{
sign_dilithium(private_key, message)
}
\arguments{
\item{private_key}{A private key produced by \code{keygen_dilithium()} to be used
for message signing.}

\item{message}{The message to be signed. Message may be interpreted in
very lax terms. Pretty much any R object can be signed, not only
character strings.}
}
\value{
A \code{pqcrypto_signature}' object.
}
\description{
Produces a digital signature of a given message.
}
\examples{
key <- keygen_dilithium(2)
important_message <- "Hello world!!"
sig <- sign_dilithium(key$private, important_message)
sig[1:10]

}
