% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_ml_dsa.R
\name{sign_ml_dsa}
\alias{sign_ml_dsa}
\title{Digital Signature - Sign ML-DSA (FIPS 204)}
\usage{
sign_ml_dsa(private_key, message, context = "")
}
\arguments{
\item{private_key}{A private key produced by \code{keygen_ml_dsa()} to be used
for message signing.}

\item{message}{The message to be signed. Message may be interpreted in
very lax terms. Pretty much any R object can be signed, not only
character strings.}

\item{context}{Optional. Defaults to the empty string, As per FIPS 204
applications may specify the use of a non-empty context string up to 255
characters.}
}
\value{
A \code{pqcrypto_signature}' object.
}
\description{
Produces a digital signature of a given message using a ML-DSA algorithm.
}
\examples{
key <- keygen_ml_dsa(2)
important_message <- "Hello world!!"
sig <- sign_ml_dsa(key$private, important_message)
sig[1:10]

}
