% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_slh_dsa.R
\name{sign_slh_dsa}
\alias{sign_slh_dsa}
\title{Digital Signature - Sign SLH-DSA (FIPS 205)}
\usage{
sign_slh_dsa(private_key, message)
}
\arguments{
\item{private_key}{A private key produced by \code{keygen_slh_dsa()} to be used
for message signing.}

\item{message}{The message to be signed. Message may be interpreted in
very lax terms. Pretty much any R object can be signed, not only
character strings.}
}
\value{
A \code{pqcrypto_signature}' object.
}
\description{
Produces a digital signature of a given message using a SLH-DSA algorithm.
}
\examples{
key <- keygen_slh_dsa()
important_message <- "Hello world!!"
signature <- sign_slh_dsa(key$private, important_message)
signature[1:16]   # first 16 bytes of the signature

}
