% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_sphincs.R
\name{sign_sphincs}
\alias{sign_sphincs}
\title{Sphincs+ Digital Signature - Sign}
\usage{
sign_sphincs(private_key, message)
}
\arguments{
\item{private_key}{A private key produced by \code{keygen_sphincs()} to be used
for message signing.}

\item{message}{The message to be signed. Message may be interpreted in
very lax terms. Pretty much any R object can be signed, not only
character strings.}
}
\value{
A \code{pqcrypto_signature}' object.
}
\description{
Produces a digital signature of a given message.
}
\examples{
key <- keygen_sphincs()
important_message <- "Hello world!!"
sig <- sign_sphincs(key$private, important_message)
sig[1:10]   # first 10 bytes of the signature

}
