% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_ml_dsa.R
\name{verify_ml_dsa}
\alias{verify_ml_dsa}
\title{Digital Signature - Verify ML-DSA (FIPS 204)}
\usage{
verify_ml_dsa(message, signature, public_key, context = "")
}
\arguments{
\item{message}{The message that has been signed.
As in \code{sign_ml_dsa()}, message may be interpreted in lax terms. It is
possible to sign any type of R objects, not only text strings.}

\item{signature}{The signature produced by \code{sign_ml_dsa()}.}

\item{public_key}{The public key created by \code{keygen_ml_dsa()} that is
paired with the private key used for signing.}

\item{context}{Optional. Defaults to the empty string, As per FIPS 204
applications may specify the use of a non-empty context string up to 255
characters.}
}
\value{
Prints the signature validation outcome and silently returns TRUE
if the signature verifies successfully or FALSE otherwise.
}
\description{
Verifies that the signature of a given message is valid.
}
\examples{
key <- keygen_ml_dsa(2)
important_message <- "Hello world!!"
signature <- sign_ml_dsa(key$private, important_message)
verify_ml_dsa(important_message, signature, key$public)

}
