% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_slh_dsa.R
\name{verify_sphincs}
\alias{verify_sphincs}
\title{Sphincs+ Digital Signature - Verify}
\usage{
verify_sphincs(message, signature, public_key)
}
\arguments{
\item{message}{The message that has been signed.
As in \code{sign_sphincs()}, message may be interpreted in lax terms. It is
possible to sign any type of R objects, not only text strings.}

\item{signature}{The signature produced by \code{sign_sphincs()}.}

\item{public_key}{The public key created by \code{keygen_sphincs()} that is
paired with the private key used for signing.}
}
\value{
Prints the signature validation outcome and silently returns TRUE
if the signature verifies successfully or FALSE otherwise.
}
\description{
Verifies that the signature of a given message is valid.
}
\examples{
key <- keygen_sphincs("sha2", 256, "fast")
important_message <- "Hello world!!"
signature <- sign_sphincs(key$private, important_message)
verify_sphincs(important_message, signature, key$public)   # Should be OK

# Try to verify a tampered message
verify_sphincs("not_the_message", signature, key$public)   # Should Fail

#' key <- keygen_slh_dsa())
important_message <- "Hello world!!"
signature <- sign_slh_dsa(key$private, important_message)
verify_sphincs(important_message, signature, key$public)
# ->
verify_slh_dsa(important_message, signature, key$public)

}
