% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_envelope.R
\name{write_envelope}
\alias{write_envelope}
\title{Envelope - Write}
\usage{
write_envelope(envelope, path = tempfile())
}
\arguments{
\item{envelope}{An envelope produced by \code{envelope_create()}.}

\item{path}{The file path. If not provided a file with a random name will
be saved in the temporary folder.}
}
\value{
Invisibly, the file path.
}
\description{
Envelope - Write
}
\examples{
key <- keygen_kyber(512)
env <- envelope_create("Very important message.", key$public)
write_envelope(env)
}
