% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_write.R
\name{write_signature}
\alias{write_signature}
\title{Digital Signature - Write}
\usage{
write_signature(signature, path = tempfile())
}
\arguments{
\item{signature}{A digital signature produced by \code{sign_dilithium()} or by
\code{sign_sphincs()}.}

\item{path}{The file path. If not provided a file with a random name will
be saved in the temporary folder.}
}
\value{
Invisibly, the file path
}
\description{
Writes a digital signature as a binary file.
}
\examples{
key <- keygen_dilithium(2)
important_message <- "Hello world!!"
signature <- sign_dilithium(key$private, important_message)
write_signature(signature)

}
