% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decap_kyber.R
\name{decap_kyber}
\alias{decap_kyber}
\title{Kyber Shared Secret - Retrieve}
\usage{
decap_kyber(encapsulation, private_key)
}
\arguments{
\item{encapsulation}{Received encapsulation object created by \code{encap_kyber()}}

\item{private_key}{Receiver's private key as generated by \code{keygen_kyber()}.}
}
\value{
The shared secret
}
\description{
Uses own private key to retrieve the secret key created by \code{encap_kyber()}.
}
\examples{
key <- keygen_kyber(512)
ss1 <- encap_kyber(key$public)  # This generates an encapsulation readable only by key$private
ss2 <- decap_kyber(ss1$encapsulation, key$private)
identical(ss1$shared_secret, ss2)

}
