% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decap_kyber.R
\name{decap_kyber}
\alias{decap_kyber}
\title{Kyber shared secret/shared key encapsulation}
\usage{
decap_kyber(encapsulation, secret_key)
}
\arguments{
\item{encapsulation}{Received encapsulation object.}

\item{secret_key}{Receiver's secret key}
}
\value{
The shared secret
}
\description{
Kyber shared secret/shared key encapsulation
}
