% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encap_kyber.R
\name{encap_kyber}
\alias{encap_kyber}
\title{Kyber shared secret/shared key encapsulation}
\usage{
encap_kyber(pub_key)
}
\arguments{
\item{pub_key}{Receiver's public key}
}
\value{
A secret key and its encapsulation to be shared with the owner of the used public key.
}
\description{
Kyber shared secret/shared key encapsulation
}
\examples{
key <- keygen_kyber(512)
s_s <- encap_kyber(key$public)  # This generates an encapsulation readable only by key$private

}
