% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encap_kyber.R
\name{encap_kyber}
\alias{encap_kyber}
\title{Kyber Shared Secret - Create}
\usage{
encap_kyber(pub_key)
}
\arguments{
\item{pub_key}{Receiver's public key as generated by \code{keygen_kyber()}.}
}
\value{
A secret key and its encapsulation to be shared with the owner of the used public key.
}
\description{
Creates a secret key and encapsulates it for secure transmission.
}
\examples{
key <- keygen_kyber(512)
ss <- encap_kyber(key$public)  # This generates an encapsulation readable only by key$private

}
