% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encapsulate_ml_kem.R
\name{encap_kyber}
\alias{encap_kyber}
\title{Kyber Shared Secret - Create}
\usage{
encap_kyber(pub_key)
}
\arguments{
\item{pub_key}{Receiver's public key as generated by \code{keygen_ml_kem()}.}
}
\value{
A list of two elements with a secret key and its encapsulation to be shared with the
owner of the used public key.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Creates a shared secret key and encapsulates it for secure transmission.
}
\examples{
key <- keygen_ml_kem(512)
ss <- encap_kyber(key$public)
# ->
ss <- encapsulate_ml_kem(key$public)

}
\keyword{internal}
