% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keygen_ml_dsa.R
\name{keygen_dilithium}
\alias{keygen_dilithium}
\title{Key-Pair Generation - Dilithium}
\usage{
keygen_dilithium(strength = 3)
}
\arguments{
\item{strength}{Type of key strength to be generated.
Different strengths are available. Use 2 for Dilithium2 (claimed strength
category 2), 3 for Dilithium3 (default, claimed strength category 3) or 5
for Dilithium5 (claimed strength category 5).
For more information about strength categories see the vignette.}
}
\value{
A \code{keypair} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
key <- keygen_dilithium()
# ->
key <- keygen_ml_dsa()
print(key)

}
\keyword{internal}
