% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keygen_sphincs.R
\name{keygen_sphincs}
\alias{keygen_sphincs}
\title{Key-Pair Generation - Sphincs+}
\usage{
keygen_sphincs(hash_type = "shake", category = 192, type = "fast")
}
\arguments{
\item{hash_type}{Type of hash to use. Accepts 'shake' (default) or 'sha2'.}

\item{category}{Security category: 128 (claimed security category 1), 192 (default, claimed
security category 2) and 256 (claimed security category 3).
For more details regarding security categories please refer to the vignette.}

\item{type}{Type of signature to produce: 'fast' (but having a larger size) or 'small'
(but being slower to compute).}
}
\value{
A \code{pqcrypto_keypair} object.
}
\description{
Key-Pair Generation - Sphincs+
}
\examples{
# Generate key with default parameters
key1 <- keygen_sphincs()
key1$algorithm
key1$param$hash

# Generate key with custom parameters
key2 <- keygen_sphincs("sha2", 128, "small")
key2$algorithm
key2$param$hash

}
