% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_ml_dsa.R
\name{sign_dilithium}
\alias{sign_dilithium}
\title{Dilithium Digital Signature - Sign}
\usage{
sign_dilithium(private_key, message)
}
\arguments{
\item{private_key}{A private key produced by \code{keygen_ml_dsa()} to be used
for message signing.}

\item{message}{The message to be signed. Message may be interpreted in
very lax terms. Pretty much any R object can be signed, not only
character strings.}
}
\value{
A \code{pqcrypto_signature}' object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Produces a digital signature of a given message using a ML-DSA algorithm.
}
\examples{
key <- keygen_ml_dsa(2)
important_message <- "Hello world!!"
sig <- sign_dilithium(key$private, important_message)
# ->
sig <- sign_ml_dsa(key$private, important_message)
sig[1:10]

}
\keyword{internal}
