% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_dillitium.R
\name{sign_dilithium}
\alias{sign_dilithium}
\title{Dilithium Digital Signature}
\usage{
sign_dilithium(private_key, msg)
}
\arguments{
\item{private_key}{The private key used for message signing.}

\item{msg}{The message to be signed.}
}
\value{
A 'pqcrypto_signature' object.
}
\description{
Dilithium Digital Signature
}
\examples{
key <- keygen_dilithium(2)
important_message <- "Hello world!!"
sig <- sign_dilithium(key$private, important_message)
sig[1:10]

}
