% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_dilithium.R
\name{verify_dilithium}
\alias{verify_dilithium}
\title{Dilithium Digital Signature - Verify}
\usage{
verify_dilithium(msg, signature, public_key)
}
\arguments{
\item{msg}{The message}

\item{signature}{The signature}

\item{public_key}{The public key}
}
\value{
TRUE if the signature verifies or FALSE otherwise.
}
\description{
Dilithium Digital Signature - Verify
}
\examples{
key <- keygen_dilithium(2)
important_message <- "Hello world!!"
sig <- sign_dilithium(key$private, important_message)
verify_dilithium(important_message, sig, key$public)

}
